Move utils source files to new directories.
