+++ packages/utils/src/Deffered/index.js
@@ -1,0 +1,2 @@
+export { default } from './Deffered';
+export * from './Deffered';
