+++ packages/utils/src/MiddlewareListener/index.js
@@ -1,0 +1,2 @@
+export { default } from './MiddlewareListener';
+export * from './MiddlewareListener';
