+++ packages/utils/src/RBAC/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './RBAC';
+export * from './RBAC';
