+++ packages/utils/src/RBAC/index.js
@@ -1,0 +1,2 @@
+export { default } from './RBAC';
+export * from './RBAC';
