+++ packages/utils/src/RBACHook/RBACHook.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { getRBAC, doesHavePermissions } from './RBAC';
+import { getRBAC, doesHavePermissions } from '../RBAC';
 
 export function usePermissions(appName, permissionsList) {
     const [ permissions, setPermissions ] = useState({ isLoading: true });
