+++ packages/utils/src/RBACHook/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './RBACHook';
+export * from './RBACHook';
