+++ packages/utils/src/RBACHook/index.js
@@ -1,0 +1,2 @@
+export { default } from './RBACHook';
+export * from './RBACHook';
