--- packages/utils/src/RBACHooks.d.ts
@@ -1,7 +1,0 @@
-import { RBAC } from './RBAC';
-
-export interface UsePermissionsState extends RBAC {
-    hasAccess: boolean;
-}
-
-declare function usePermissions(appName: string, permissionsList: string[]): UsePermissionsState
