+++ packages/utils/src/ReducerRegistry/index.js
@@ -1,0 +1,2 @@
+export { default } from './ReducerRegistry';
+export * from './ReducerRegistry';
