+++ packages/utils/src/Registry/index.js
@@ -1,0 +1,2 @@
+export { default } from './Registry';
+export * from './Registry';
