+++ packages/utils/src/RouterParams/index.js
@@ -1,0 +1,2 @@
+export { default } from './RouterParams';
+export * from './RouterParams';
