+++ packages/utils/src/RowLoader/index.js
@@ -1,0 +1,2 @@
+export { default } from './RowLoader';
+export * from './RowLoader';
