--- packages/utils/src/Utilities.scss
@@ -1,1 +1,0 @@
-@import '_all';
