+++ packages/utils/src/debounce/index.js
@@ -1,0 +1,2 @@
+export { default } from './debounce';
+export * from './debounce';
