+++ packages/utils/src/helpers/helpers.test.js
@@ -1,10 +1,5 @@
-import React from 'react';
-import { mount, shallow } from 'enzyme';
-import toJson from 'enzyme-to-json';
 import {
-    RowLoader,
     processDate,
-    parseCvssScore,
     downloadFile,
     mergeArraysByKey,
     getBaseName,
@@ -68,50 +63,6 @@ describe('downloadFile', () => {
     });
 });
 
-describe('parseCvssScore', () => {
-    describe('no cvss', () => {
-        const result = parseCvssScore();
-        const wrapper = mount(result);
-        it('should render N/A in span', () => {
-            expect(toJson(wrapper.find('span'))).toMatchSnapshot();
-        });
-
-        it('should render correct tooltip content', () => {
-            expect(toJson(wrapper.find('TooltipContent'))).toMatchSnapshot();
-        });
-    });
-
-    it('cvssV3', () => {
-        const result = parseCvssScore(1, 2);
-        const wrapper = mount(<React.Fragment>{ result }</React.Fragment>);
-        expect(toJson(wrapper)).toMatchSnapshot();
-    });
-
-    describe('cvssV2', () => {
-        const result = parseCvssScore(2);
-        const wrapper = mount(result);
-        it('should render value in span', () => {
-            expect(toJson(wrapper.find('span'))).toMatchSnapshot();
-        });
-
-        it('should render correct tooltip content', () => {
-            expect(toJson(wrapper.find('TooltipContent'))).toMatchSnapshot();
-        });
-    });
-
-    describe('cvssV2 with labels', () => {
-        const result = parseCvssScore(2, 0, true);
-        const wrapper = mount(result);
-        it('should render value in span', () => {
-            expect(toJson(wrapper.find('span'))).toMatchSnapshot();
-        });
-
-        it('should render correct tooltip content', () => {
-            expect(toJson(wrapper.find('TooltipContent'))).toMatchSnapshot();
-        });
-    });
-});
-
 describe('processDate', () => {
     it('should return parsed date', () => {
         expect(processDate(1558530000000)).toBe('22 May 2019');
@@ -124,13 +75,6 @@ describe('processDate', () => {
     it('should render parse date correctly', () => {
         expect(processDate(1557000000000)).toBe('04 May 2019');
 
-    });
-});
-
-describe('RowLoader', () => {
-    it('should render correctly', () => {
-        const wrapper = shallow(<RowLoader />);
-        expect(toJson(wrapper)).toMatchSnapshot();
     });
 });
