+++ packages/utils/src/interceptors/index.js
@@ -1,0 +1,2 @@
+export { default } from './interceptors';
+export * from './interceptors';
