+++ packages/utils/src/inventoryDependencies/index.js
@@ -1,0 +1,5 @@
+/**
+ * @deprecated this modules are deprectated and no longer required in chrome 2.0
+ */
+export { default } from './inventoryDependencies';
+export * from './inventoryDependencies';
