+++ packages/utils/src/mergeMessages/index.js
@@ -1,0 +1,1 @@
+require('./mergeMessages');
