+++ packages/utils/src/parseCvssScore/index.js
@@ -1,0 +1,2 @@
+export { default } from './parseCvssScore';
+export * from './parseCvssScore';
