+++ packages/utils/src/useInventory/index.js
@@ -1,0 +1,2 @@
+export { default } from './useInventory';
+export * from './useInventory';
